Refactor manual, client imagediff
