+++ test/integration/tools/manual/client/frame.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>VizzuTest</title>
+    <title>Vizzu library test</title>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
   </head>
@@ -19,6 +19,6 @@
       id="vizzuCanvas"
       style="width: 640px; height: 360px; border: 1px solid gray"
     ></canvas>
-    <script id="vizzuTest" src="frame.js"></script>
+    <script src="frame.js"></script>
   </body>
 </html>
