+++ test/integration/tools/manual/client/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Vizzu library test - manual check</title>
+    <title>Vizzu library test - manual</title>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
   </head>
